Changed average ratings to floats

+++ db/schema.rb
@@ -49,10 +49,10 @@ ActiveRecord::Schema.define(version: 20180914231206) do
     t.string   "image"
     t.integer  "price_cents"
     t.integer  "quantity"
-    t.datetime "created_at",                 null: false
-    t.datetime "updated_at",                 null: false
+    t.datetime "created_at",                   null: false
+    t.datetime "updated_at",                   null: false
     t.integer  "category_id"
-    t.integer  "average_rating", default: 0
+    t.float    "average_rating", default: 0.0
   end
 
   add_index "products", ["category_id"], name: "index_products_on_category_id", using: :btree
